canandmessage/pyproject.toml
[project]
name = "canandmessage"
version = "4.6.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jinja2>=3.1.5",
    "pyserial>=3.5",
    "python-can[gs-usb]>=4.5.0",
    "pyusb>=1.2.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "websockets>=14.2",
]
